Entity ID types do not match types in database definition
As Liquibase is used, JPA does not manage database definition.
Within Liquibase changelog, only Anchor and Fragment have BIGINT (Long)
IDs. Other entities have INTEGER IDs.

- Change AnchorEntity/AnchorRepository ID type to Long
- Change YangResourceEntity/YangResourceRepository and
  ModuleReferenceRepository ID to Integer
- Modify datatypes of foreign key references in liquibase changelog to
  use correct types for dataspace_id, schema_set_id, yang_resource_id
- Update test liquibase changelog

Issue-ID: CPS-1676

+++ cps-ri/src/main/java/org/onap/cps/spi/impl/CpsDataPersistenceServiceImpl.java
@@ -346,10 +346,10 @@ public class CpsDataPersistenceServiceImpl implements CpsDataPersistenceService 
             fragmentRepository.findExtractsWithDescendantsByIds(fragmentEntityIds, fetchDescendantsOption.getDepth());
 
         if (anchorEntity == ALL_ANCHORS) {
-            final Collection<Integer> anchorIds = fragmentExtracts.stream()
+            final Collection<Long> anchorIds = fragmentExtracts.stream()
                 .map(FragmentExtract::getAnchorId).collect(Collectors.toSet());
             final List<AnchorEntity> anchorEntities = anchorRepository.findAllById(anchorIds);
-            final Map<Integer, AnchorEntity> anchorEntityPerId = anchorEntities.stream()
+            final Map<Long, AnchorEntity> anchorEntityPerId = anchorEntities.stream()
                 .collect(Collectors.toMap(AnchorEntity::getId, Function.identity()));
             return FragmentEntityArranger.toFragmentEntityTreesAcrossAnchors(anchorEntityPerId, fragmentExtracts);
         }
